Move document tab and languages to their own file.
