+++ crates/planner_gui/src/main.rs
@@ -3,102 +3,26 @@
     windows_subsystem = "windows"
 )]
 
-use std::str::FromStr;
-use std::sync::{Mutex, MutexGuard};
-use tracing::{info, debug};
+use tracing::info;
 use freya::prelude::*;
 use dioxus_router::prelude::{Outlet, Routable, Router, use_route};
-use dioxus_sdk::{
-    i18n::{
-        use_i18,
-        use_init_i18n,
-        Language,
-    },
-    translate,
-};
-use unic_langid::LanguageIdentifier;
 use tracing_subscriber::{EnvFilter, fmt};
 use tracing_subscriber::layer::SubscriberExt;
 use tracing_subscriber::util::SubscriberInitExt;
-
 use planner_app::ViewModel;
-use crate::app_core::CoreService;
+
 use crate::toolbar::ToolBar;
 
 mod app_core;
 mod toolbar;
-
-static LANGUAGES: Mutex<Vec<LanguagePair>> = Mutex::new(Vec::new());
-static SELECTED_LANGUAGE: Mutex<Option<LanguagePair>> = Mutex::new(None);
-
-// TODO make loading languages dynamic so that translators don't have to re-compile to test.
-//      the code is prepared for this by using the two mutexes above instead of using
-//      static constants.
-fn initialise_languages() {
-
-    let languages: Vec<LanguagePair> = vec![
-        LanguagePair { code: "en-US".to_string(), name: "English (United-States)".to_string() },
-        LanguagePair { code: "es-ES".to_string(), name: "Español (España)".to_string() },
-    ];
-    
-    let first_language: &LanguagePair = languages.first().unwrap();
-    let first_language_identifier: LanguageIdentifier = first_language.code.parse().unwrap();
-    
-    use_init_i18n(first_language_identifier.clone(), first_language_identifier, || {
-        languages.iter().map(|LanguagePair { code, name: _name }|{
-            match code.as_str() {
-                "en-US" => Language::from_str(EN_US).unwrap(),
-                "es-ES" => Language::from_str(ES_ES).unwrap(),
-                _ => panic!()
-            }
-        }).collect()
-    });
-
-    let mut guard = SELECTED_LANGUAGE.lock().unwrap();
-    (*guard).replace(first_language.clone());
-
-    let mut guard = LANGUAGES.lock().unwrap();
-    (*guard).extend(languages);
-    
-}
-
-fn change_language(language_pair: &LanguagePair) {
-
-    let mut i18n = use_i18();
-
-    let mut guard = SELECTED_LANGUAGE.lock().unwrap();
-    (*guard).replace(language_pair.clone());
-
-    i18n.set_language(language_pair.code.parse().unwrap());
-}
-
-fn languages() -> MutexGuard<'static, Vec<LanguagePair>> {
-    
-    let guard = LANGUAGES.lock().expect("not locked");
-    
-    guard
-}
-
-#[derive(Clone)]
-struct LanguagePair {
-    code: String,
-    name: String,
-}
-
-// FIXME avoid cloning, return some reference instead, but how!
-fn selected_language() -> LanguagePair {
-
-    let guard = SELECTED_LANGUAGE.lock().expect("not locked");
-
-    guard.as_ref().unwrap().clone()
-}
+mod document_tab;
+mod languages;
 
 fn app() -> Element {
 
-    initialise_languages();
+    languages::initialise();
 
-    change_language(&selected_language());
-
+    languages::change(&languages::selected());
 
     rsx!(
         rect {
@@ -257,18 +181,6 @@ fn PageNotFound() -> Element {
         }
     )
 }
-fn use_change_language() -> Signal<Box<(impl FnMut(LanguagePair) + 'static)>> {
-    let mut i18n = use_i18();
-
-    let closure = move |language_pair: LanguagePair| {
-        let mut guard = SELECTED_LANGUAGE.lock().unwrap();
-
-        i18n.set_language(language_pair.code.parse().unwrap());
-        (*guard).replace(language_pair);
-    };
-    
-    use_signal(move || Box::new(closure))
-}
 
 #[allow(non_snake_case)]
 fn DocumentRouteLayout() -> Element {
@@ -283,7 +195,7 @@ fn DocumentRouteLayout() -> Element {
 #[allow(non_snake_case)]
 fn DocumentLayout() -> Element {
 
-    let view = use_signal(ViewModel::default);
+    let _view = use_signal(ViewModel::default);
 
     // TODO get the current document?
     let document_path: DocumentRoute = use_route();
@@ -317,10 +229,6 @@ fn DocumentLayout() -> Element {
 }
 
 static ARIMO_NERD_FONT: &[u8] = include_bytes!("../assets/fonts/ArimoNerdFont-Regular.ttf");
-
-
-static EN_US: &str = include_str!("../assets/i18n/en-US.json");
-static ES_ES: &str = include_str!("../assets/i18n/es-ES.json");
 
 fn main() {
